Add optional output compression
To allow large build directories to be stored, given that they need to
be put into a tarball somewhere before they can be uploaded, allow them
to be compressed with the user's choice of compression algorithm first.
Normally we can just let GitHub worry about compression, but given this
file needs to be stored, if only temporarily, on disk, that may not be
an option.